Moved Input and Restart file reading from main to input.c
