+++ src/input.c
@@ -31,3 +31,78 @@ int get_a_line(FILE *fp, char *buf)
     }
     return 0;
 }
+
+
+int readInputRest(mdsys_t *sys , char * restfile, char * trajfile, char * ergfile, char * line, int *nprint){
+ 
+     int i;
+     FILE *fp;
+ 
+     /* read input file */
+     if(get_a_line(stdin,line)) return 1;
+     sys->natoms=atoi(line);
+     if(get_a_line(stdin,line)) return 1;
+     sys->mass=atof(line);
+     if(get_a_line(stdin,line)) return 1;
+     sys->epsilon=atof(line);
+     if(get_a_line(stdin,line)) return 1;
+     sys->sigma=atof(line);
+     if(get_a_line(stdin,line)) return 1;
+     sys->rcut=atof(line);
+     if(get_a_line(stdin,line)) return 1;
+     sys->box=atof(line);
+     if(get_a_line(stdin,restfile)) return 1;
+     if(get_a_line(stdin,trajfile)) return 1;
+     if(get_a_line(stdin,ergfile)) return 1;
+     if(get_a_line(stdin,line)) return 1;
+     sys->nsteps=atoi(line);
+     if(get_a_line(stdin,line)) return 1;
+     sys->dt=atof(line);
+     if(get_a_line(stdin,line)) return 1;
+     *nprint=atoi(line);
+ 
+     /* allocate memory */
+     sys->rx=(double *)malloc(sys->natoms*sizeof(double));
+     sys->ry=(double *)malloc(sys->natoms*sizeof(double));
+     sys->rz=(double *)malloc(sys->natoms*sizeof(double));
+     sys->vx=(double *)malloc(sys->natoms*sizeof(double));
+     sys->vy=(double *)malloc(sys->natoms*sizeof(double));
+     sys->vz=(double *)malloc(sys->natoms*sizeof(double));
+     sys->fx=(double *)malloc(sys->natoms*sizeof(double));
+     sys->fy=(double *)malloc(sys->natoms*sizeof(double));
+     sys->fz=(double *)malloc(sys->natoms*sizeof(double));
+
+/* read restart */
+     fp=fopen(restfile,"r");
+     if(fp) {
+       int err;
+         for (i=0; i<sys->natoms; ++i) {
+             err = fscanf(fp,"%lf%lf%lf",sys->rx+i, sys->ry+i, sys->rz+i);
+         }
+         for (i=0; i<sys->natoms; ++i) {
+             err = fscanf(fp,"%lf%lf%lf",sys->vx+i, sys->vy+i, sys->vz+i);
+         }
+         fclose(fp);
+         azzero(sys->fx, sys->natoms);
+         azzero(sys->fy, sys->natoms);
+         azzero(sys->fz, sys->natoms);
+     (void)err;
+     // if err != 0 return -1;
+     } else {
+         perror("cannot read restart file");
+         return 3;
+     }
+ 
+     return 0;
+ }
+
+
+
+
+
+
+
+
+
+
+
